+++ client/package.json
@@ -8,7 +8,7 @@
 		"build": "react-scripts build",
 		"predeploy": "npm run build",
 		"deploy": "vercel deploy --prebuilt",
-		"compile": "graphql-codegen",
+		"gen": "graphql-codegen",
 		"watch": "graphql-codegen -w"
 	},
 	"dependencies": {
