+++ server/src/resolvers.ts
@@ -1,11 +1,11 @@
 import { Query } from './resolvers/query';
-import { Space } from './resolvers/space';
-import { Map } from './resolvers/map';
+import { Space } from './resolvers/tsd-map/space';
+import { Map } from './resolvers/tsd-map/map';
 import { Mutation } from './resolvers/mutation';
-import { LightLevel } from './resolvers/light-level';
-import { MapSettings } from './resolvers/map-settings';
-import { User } from './resolvers/user';
-import { SpaceType } from './resolvers/space-type';
+import { LightLevel } from './resolvers/tsd-map/light-level';
+import { MapSettings } from './resolvers/tsd-map/map-settings';
+import { User } from './resolvers/tsd-map/user';
+import { SpaceType } from './resolvers/tsd-map/space-type';
 import { RetailAccount } from './resolvers/retail/retail-account';
 
 export const resolvers = {
