+++ server/src/resolvers/tsd-map/light-level.ts
@@ -1,0 +1,5 @@
+import { DBLightLevel } from '@/types/tsd-map/light-level';
+
+export const LightLevel = {
+  movementPoints: (lightLevel: DBLightLevel) => lightLevel.movement_points,
+};
