+++ server/src/resolvers/tsd-map/space-type.ts
@@ -1,0 +1,5 @@
+import { DBSpaceType } from '@/types/tsd-map/space-type';
+
+export const SpaceType = {
+  iconUrl: (spaceType: DBSpaceType) => spaceType.icon_url,
+};
