+++ server/src/resolvers/tsd-map/user.ts
@@ -1,0 +1,5 @@
+import { DBUser } from '@/types/user/user';
+
+export const User = {
+  viewedSetup: (user: DBUser) => user.viewed_setup,
+};
