--- server/src/types/light-level.ts
@@ -1,11 +1,0 @@
-export interface LightLevel {
-	id: string;
-	name: string;
-	movementCost: number;
-}
-
-export interface DBLightLevel {
-	id: string;
-	name: string;
-	movement_points: number;
-}
