--- server/src/types/space-group.ts
@@ -1,12 +1,0 @@
-export interface SpaceGroup {
-	id: string;
-	name: string;
-	prefix: string;
-}
-
-export interface DBSpaceGroup {
-	id: string;
-	map_id: string;
-	name: string;
-	prefix: string;
-}
