+++ server/src/types/tsd-map/light-level.ts
@@ -1,0 +1,11 @@
+export interface LightLevel {
+  id: string;
+  name: string;
+  movementCost: number;
+}
+
+export interface DBLightLevel {
+  id: string;
+  name: string;
+  movement_points: number;
+}
