+++ server/src/types/tsd-map/map-settings.ts
@@ -1,0 +1,23 @@
+export interface MapSettings {
+  backgroundImageUrl: string;
+  spaceColor: string;
+  horizontalSpacing: number;
+  verticalSpacing: number;
+  indent: number;
+  paddingX: number;
+  paddingY: number;
+  spaceRadius: number;
+}
+
+export interface DBMapSettings {
+  id: string;
+  map_id: string;
+  background_image_url: string;
+  space_color: string;
+  horizontal_spacing: number;
+  vertical_spacing: number;
+  indent: number;
+  padding_x: number;
+  padding_y: number;
+  space_radius: number;
+}
