+++ server/src/types/tsd-map/map.ts
@@ -3,14 +3,14 @@ import { DBSpace } from './space';
 import { SpaceGroup } from './space-group';
 
 export interface SDMap {
-	id: string;
-	title: string;
-	settings: MapSettings;
-	spaces: DBSpace[] | Promise<DBSpace[]>;
-	spaceGroups: SpaceGroup[] | Promise<SpaceGroup[]>;
+  id: string;
+  title: string;
+  settings: MapSettings;
+  spaces: DBSpace[] | Promise<DBSpace[]>;
+  spaceGroups: SpaceGroup[] | Promise<SpaceGroup[]>;
 }
 
 export interface DBMap {
-	id: string;
-	name: string;
+  id: string;
+  name: string;
 }
