+++ server/src/types/tsd-map/object.ts
@@ -1,3 +1,3 @@
 export enum ObjectType {
-	SPACE = 'SPACE',
+  SPACE = 'SPACE',
 }
