+++ server/src/types/tsd-map/space-group.ts
@@ -1,0 +1,12 @@
+export interface SpaceGroup {
+  id: string;
+  name: string;
+  prefix: string;
+}
+
+export interface DBSpaceGroup {
+  id: string;
+  map_id: string;
+  name: string;
+  prefix: string;
+}
