+++ server/src/types/tsd-map/space-type.ts
@@ -1,0 +1,15 @@
+export interface SpaceType {
+  id: number;
+  name: string;
+  iconUrl?: string;
+  description: string;
+  color?: string;
+}
+
+export interface DBSpaceType {
+  id: string;
+  name: string;
+  icon_url?: string;
+  description: string;
+  color: string;
+}
