--- server/src/types/user.ts
@@ -1,13 +1,0 @@
-export interface User {
-	id: string;
-	clerkId: string;
-	email: string;
-	viewedSetup: boolean;
-}
-
-export interface DBUser {
-	id: string;
-	clerk_id: string;
-	email: string;
-	viewed_setup: boolean;
-}
