+++ server/src/types/user/user.ts
@@ -1,0 +1,13 @@
+export interface User {
+  id: string;
+  clerkId: string;
+  email: string;
+  viewedSetup: boolean;
+}
+
+export interface DBUser {
+  id: string;
+  clerk_id: string;
+  email: string;
+  viewed_setup: boolean;
+}
